Add files via upload
Diabetes is a chronic illness characterized by elevated levels of blood glucose, accompanied by disturbed metabolism of fats and proteins. The prevalence of diabetes is increasing worldwide. The International Diabetes Federation (IDF) estimates that 536.6 million people are living with diabetes (diagnosed or undiagnosed) in 2021, and this number is projected to increase by 46%, reaching 783.2 million by 2045. 
The health experts and world health organization revealed that the incidence of diabetes chronic disease and also other non-communicable diseases are increasing day by day all over the world. Monitoring and analysis tools for diabetes care often include information visualizations to help people make sense of the complex data, The use of pandas, numpy, seaborn and matplotlib allows for efficient data manipulation, analysis and data visualization. The dataset used in the project is a self-report data and is optained from The Behavioral Risk Factor Surveillance System (BRFSS) which is a health-related telephone survey that is collected annually by the CDC. The result presented that the body mass index (BMI), high blood pressure, high cholesterol and age has a significance impact of the diabetes disease and the increase or growth of the mentioned factors can cause the individual to be diagnosed with diabetes cases while Income, Education and physical Activity have a negative impact or negative relation with Diabetes disease